move RelayContext to modern
Summary: Modern uses this all over the place. `/classic/` should only contain things we'll eventually want to delete and `/modern/` should ideally never require into `/classic/`.

Reviewed By: jstejada

Differential Revision: D9944219

fbshipit-source-id: e2bef61ef7605b69c8f7320933d5c94c9da811f4
