+++ packages/react-relay/modern/isRelayEnvironment.js
